Blend color with sidebar color instead of replacing it
Added support to blend the colors with the sidebar background color instead of replacing it. This drastically improves the appearance with the default Xcode theme.

Options have been added to the preferences to enable/disable blending, as well as to set the blend factor.

The default colors now are opaque, as blending factor is used to control how weak or strong the color appears.